requirements.txt
streamlit==1.25.0
python-dotenv==1.0.0
openai==1.3.7
langchain==0.0.353
pydantic==1.10.13
PyMuPDF==1.22.5
pdfplumber==0.9.0
pytesseract==0.3.10
Pillow==9.5.0